backend/requirements.txt
fastapi
uvicorn
pydantic
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-google-genai==2.0.0
faiss-cpu
python-dotenv
sqlalchemy
sentence-transformers

